if else
This program is a simple guessing game. The computer is to generate a
random number between 1 and 20. The user is given up to five tries to
guess the exact number. After each guess, you are to tell the user if the
guessed number is greater than, less than, or equal to the random number. If it is equal, no more guesses should he made. If the user hasn’
t
guessed the number after five tries, display the number with a message
that the user should know it by now and terminate the game.